requirements_move_data.txt
# Function dependencies, for example:
# package>=version
PyMySQL
google-cloud-storage